requirements.txt
numpy
pyyaml
transforms3d
pyrealsense2
open3d
opencv-python